pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-test-kit"
version = "0.1.0"
description = "Reusable testing framework for LLM agents: fluent session API, deterministic + LLM-as-Judge assertions, pytest integration."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "KASKO Team" }]

dependencies = [
    "pytest>=8.0",
    "requests>=2.31",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
deepeval = [
    "deepeval>=2.0",
]
gigachat = [
    "langchain-gigachat>=0.3",
    "langchain-core>=0.3",
]
openai = [
    "openai>=1.0",
]
anthropic = [
    "anthropic>=0.30",
]
allure = [
    "allure-pytest>=2.13",
]
all = [
    "agent-test-kit[allure,deepeval,gigachat,openai,anthropic]",
]
dev = [
    "allure-pytest>=2.13",
    "pytest>=8.0",
    "ruff>=0.4",
]

[project.entry-points."pytest11"]
agent_test_kit = "agent_test_kit.pytest_plugin"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_test_kit"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "judge: tests that use LLM-as-Judge style evaluation",
    "slow: slower-running tests such as repeated runs or concurrency checks",
]
